fix alertmanager pod labels and pod selector labels
previously we had used label keys and values, and
annotation keys and values that were actually
reserved by the alertmanager operator i.e the
prometheus operator which has the alertmanager
operator inside it. now we ensure that we don't
use any reserved labels and any reserved
annotations. we now have a separate set of labels
for all resources and a separate set of labels for
alertmanager pods

Signed-off-by: Karuppiah Natarajan <karuppiah7890@gmail.com>